COM-2247 requests access to the camera when user starts timeStamp
